requirements.txt
pytest>=8.0
allure-pytest>=2.13
requests>=2.31
pydantic>=2.0
pydantic-settings>=2.0
pytest-xdist>=3.5
sqlalchemy>=2.0
psycopg2-binary>=2.9
alembic>=1.13
cryptography>=42.0
pyotp>=2.9
